Added files via upload
A sample CTH simulation of an impact between rock/ice exoplanets, each composed of a mixture of ice and dunite, with M_T=0.35M_E, gamma=0.14, impact angle=50 degrees, and v_{imp}/v_{esc}=0.87.  The resulting disk is 30 % the mass of Jupiter's moon Ganymede.  The estimated satellite density is 1.19 g/cm^3, yielding a maximum satellite radius R~2000 km, in between the sizes of Earth's Moon and the planet Mercury.  An animation of this figure may be viewed online: https://youtu.be/JIK9b_Qb37Q.